+++ lib/src/network/winr_api.dart
@@ -253,7 +253,7 @@ class WINRRequestDefaults {
   WINRRequestDefaults._();
 
   static String platformOS = 'iOS';
-  static String sdkVersion = 'v2.0.0';
+  static String sdkVersion = '2.6.2';
 }
 
 /// Response from claiming daily entries (mirrors iOS